RuntimeError in multiprocessing
On Windows the multiprocessing library relies on the spawn backend will import (i.e. execute) the main module at start. While this is (negligibly) slower, it is safer. An if __name__ == '__main__': guard in the external command will avoid creating itself recursively, resolving the RuntimeError. On Linux, the code will still default to the "fork" backend, which is not available on Windows.

This appears to be correct in the 2D and 3D xanes module, but without test data I am unable to verify.