Resolve ValueError: need at most 63 handles on Windows
On Windows hosts with 64 logical cores or more, multiprocessing library will error out because python will exceed the Windows limitation of _winapi.WaitForMultipleObjects. When a host machine with an invalid configuration (64 logical processors and windows OS) is detected, limit processes created down to 40. While resources online recommend 60 or 61 to account for any existing python processes, occasionally, I've experimentally found that Tomopy(?) will not always clean up, so a buffer of 20 threads was given to have a consistently stable environment for batch jobs. As to the performance hit, I've found it to not be significant as the processor is still nearly fully utilized (All 32 physical cores are fully saturated on our AMD host, so therefore all 64 logical cores report a near 100% utilization) (This may differ between Intel's hyperthreading, and AMD's SMT, but that is out of the scope of this fix).